BHYVE: OS-7843 viona could be split up
commit 9e88ade90f
